Rename unit test file
